requirements.txt
django>=4.2
djangorestframework
django-cors-headers
python-dotenv
langchain
langchain-groq
langchain-text-splitters
langchain-community
langchain-huggingface
langchain-qdrant
faiss-cpu
pypdf
python-docx
sentence-transformers
pymongo[srv]
qdrant-client
gunicorn
whitenoise
